% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin(iter, n, p)}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations to be run and simulated}

\item{n}{The simulated sample size}

\item{p}{The probability of success}
}
\value{
A colored barplot of the computed values
}
\description{
Binomial Experiment Simulation Function.
}
\details{
This function is designed to designed to simulate a binomial experiment using the given parameters. These
parameters include the number of iterations that are to be run, the desired sample size to be simulated,
and a given probability.
}
\examples{
mybin(iter=100,n=10, p=0.7)
mybin(iter=1500,n=60, p=0.8)
mybin(iter=2800,n=15, p=0.5)
}
