% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{myci(x, conf)}
\usage{
myci(x, conf = 0.95)
}
\arguments{
\item{x}{The data to be used to compute the confidence interval}

\item{conf}{The confidence interval level as a decimal (i.e. 95 = .95). By default this function computes 95 unless specified otherwise}
}
\value{
A concatenated output containing the Lower and Upper bounds of the confidence interval. The format is as follows: L , U
}
\description{
Confidence Interval Computation Function.
}
\details{
This function is designed to take in a data set and compute a confidence interval to the
provided confidence level.
}
\examples{
myci(x=data, conf=.8)
myci(x=c(12, 45, 17, 39, 29), conf =.9)
}
