% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu(n, iter, a, b)}
\usage{
mycltu(n, iter, a, b)
}
\arguments{
\item{n}{The desired sample size}

\item{iter}{The number of iterations to run to compute the random sample}

\item{a}{The a value used to calculate the mean and variance for the theoretical curve}

\item{b}{The b value used to calculate the mean and variance for the theoretical curve}
}
\value{
A histogram with the sample mean, curve made from the sample distribution computed, and a
theoretical normal curve.
}
\description{
Uniform Central Limit Theorem Function.
}
\details{
This function takes in various parameters to compute the Central Limit Theorem based on a computed
uniform random sample. This information is then overlaid graphically on a histogram, along with the sampling
distribution that is then returned out of the function.
}
\examples{
mycltu(3, 10000, 0, 10)
mycltu(5, 10000, 0, 10)
mycltu(1, 10000, 0, 10)
}
