% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myreadin.R
\name{myreadin}
\alias{myreadin}
\title{myreadin(filename)}
\usage{
myreadin(filename)
}
\arguments{
\item{filename}{The name of the file to be read in, excluding the extension. Must be either CSV or XLS}
}
\value{
A data frame with data from the file that was read in
}
\description{
Reading in Data Files Function.
}
\details{
This function is meant to take the small annoyance out of having to load in either a CSV or XLS and
does the work for you. You pass in the file name WITHOUT the extension and the function will check if
it exists, whether its a CSV or XLS, read it in, and return the data to be caught in a variable. If the
file doesn't exist/can't be found, it spits out a File Not Found message with a sad face.
Make sure the data is located in the \strong{\emph{caller file's}} current working directory.
}
\examples{
myreadin('CIVILSAL')
myreadin('NZBIRDS')
myreadin('DDT')
}
